requirements.txt
flet==0.19.0
icecream==2.1.3
PyGithub==2.2.0
requests==2.31.0
